drivers: iommu: Fix Cert-C error handling bug
In function smmu_context_filter_show casting find_next_bit function
from unsigned long to int without checking its value may result in lost
or misinterpreted data. Fix this by changing data type of variable.

CID 490421
Bug 3512545

Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2736802
GVS: Gerrit_Virtual_Submit
Tested-by: Sachin Nikam <snikam@nvidia.com>

+++ drivers/iommu/arm-smmu-debug.c
@@ -150,13 +150,13 @@ static int smmu_context_filter_show(struct seq_file *s, void *unused)
 {
 	struct smmu_debugfs_info *smmu_dfs = s->private;
 	unsigned long *bitmap = smmu_dfs->context_filter;
-	int idx = 0;
+	unsigned long idx = 0;
 
 	while (1) {
 		idx = find_next_bit(bitmap, smmu_dfs->max_cbs, idx);
 		if (idx >= smmu_dfs->num_context_banks)
 			break;
-		seq_printf(s, "%d,", idx);
+		seq_printf(s, "%lu,", idx);
 		idx++;
 	}
 	seq_putc(s, '\n');
